start docker migration
